fix: waveform index bug

+++ pt-client/src/views/timeline.rs
@@ -134,10 +134,10 @@ fn generate_focii(tracks: &HashMap<u16, Track>,
                     state.sample_rate,
                     state.tempo,
                     state.zoom);
-                let in_x = offset_in - state.scroll_x;
+                let in_x = offset_in as i16 - state.scroll_x as i16;
                 if in_x >= 0 {
                     write!(out, "{} <<", cursor::Goto(
-                        window.x + REGIONS_X + in_x - 3, 
+                        window.x + REGIONS_X + in_x as u16 - 3, 
                         window.y + TIMELINE_Y)
                     ).unwrap()
                 }
